Add env vars and exit status exploration
As regards to environment variables, sketch code that returns, more
or less (need to double check edge cases, specific function
behaviour, special parameter protection, ...) what is expected of
the built-ins:
- export with no options
- unset with no options
- env with no options or arguments.

As also regards to environment variables, sketch code  that:
Takes in a string with environment variables specified ($NAME)
Returns  an array of strings with the environment variables
replaced with their respective values, if set, or an empty string,
if not set.
Known limitations:
- Replaces every '$' character. So, as of right now, it is only a function
that can only be used to expand a string where every '$' signals an environment
variable.
- As of right now it does not expand '$?'. '$?' is a special parameter that
stands for the latest exit status. Though this can easily be added to this
function. The only question is whether we want to think of '$?' as an
environment variable and/or as a simple element that makes up part of
a possible 'global' static struct that holds all the information we decide
to keep in there (environment variables, ...).

As regards to the exit status, sketch a function that updates a static
struct element designed to hold the latest exit status.
Still trying to understand why an exit status code of 131 should be handled
differently.