Fix export bug
Example: export NAME should not generate an error message,
eventhough it does not update the environment variables' linked
list. Such an example now behaves as expected.

Use of get_env_key() was passing a NULL value - as it wasn't able
to find a '=' character - to invalid_key(). Hence triggering an
error message.