Fix standalone redirections
For standalone redirections only heredocs seem to be really
generated. The current commit tries to mimic this. For standalone
non heredocs redirections no dup2() call is made. Only: file
descriptors are opened and closed. This should ensure errors are
triggered for invalid file descriptors on stdin redirections and
files are created for stdout redirections.